refactor: move DatabaseError and ComponentError to misc.exceptions
