+++ flood_adapt/__init__.py
@@ -4,8 +4,8 @@ __version__ = "0.3.7"
 from flood_adapt import adapter, dbs_classes, objects
 from flood_adapt.config.config import Settings
 from flood_adapt.config.site import Site
-from flood_adapt.exceptions import ComponentError, DatabaseError, FloodAdaptError
 from flood_adapt.flood_adapt import FloodAdapt
+from flood_adapt.misc.exceptions import ComponentError, DatabaseError, FloodAdaptError
 from flood_adapt.misc.log import FloodAdaptLogging
 from flood_adapt.objects.forcing import unit_system
