+++ flood_adapt/dbs_classes/dbs_measure.py
@@ -4,7 +4,7 @@ from typing import Any
 import geopandas as gpd
 
 from flood_adapt.dbs_classes.dbs_template import DbsTemplate
-from flood_adapt.exceptions import DatabaseError
+from flood_adapt.misc.exceptions import DatabaseError
 from flood_adapt.misc.utils import resolve_filepath
 from flood_adapt.objects.measures.measure_factory import MeasureFactory
 from flood_adapt.objects.measures.measures import Measure
