+++ flood_adapt/dbs_classes/dbs_static.py
@@ -10,7 +10,7 @@ from flood_adapt.adapter.sfincs_adapter import SfincsAdapter
 from flood_adapt.config.config import Settings
 from flood_adapt.dbs_classes.interface.database import IDatabase
 from flood_adapt.dbs_classes.interface.static import IDbsStatic
-from flood_adapt.exceptions import DatabaseError
+from flood_adapt.misc.exceptions import DatabaseError
 
 
 def cache_method_wrapper(func: Callable) -> Callable:
