+++ flood_adapt/misc/exceptions.py
