+++ tests/test_dbs_classes/test_database.py
@@ -7,7 +7,7 @@ import pytest
 from flood_adapt.config.config import Settings
 from flood_adapt.config.site import Site
 from flood_adapt.dbs_classes.database import Database
-from flood_adapt.exceptions import DatabaseError
+from flood_adapt.misc.exceptions import DatabaseError
 from flood_adapt.workflows.benefit_runner import Benefit, BenefitRunner
